Optimized check of configuration file

+++ tests/test_base.py
@@ -15,15 +15,12 @@ TASKFILE: ~/
 '''
         self.test_conf_file = os.path.expanduser(os.path.join('~', '.letsdo'))
 
-        try:
+        if os.path.exists(self.test_conf_file):
             with open(self.test_conf_file, 'r') as f:
                 self.backup = f.read()
 
             with open(self.test_conf_file, 'w') as f:
                 f.write(test_configuration)
-        except IOError:
-            # Nothing to backup
-            pass
 
         self.conf = Configuration()
 
@@ -33,8 +30,9 @@ TASKFILE: ~/
             os.remove(self.conf.task_filename)
 
     def tearDown(self):
-        with open(self.test_conf_file, 'w') as f:
-            f.write(self.backup)
+        if os.path.exists(self.test_conf_file):
+            with open(self.test_conf_file, 'w') as f:
+                f.write(self.backup)
         if os.path.exists(self.conf.data_filename):
             os.remove(self.conf.data_filename)
         if os.path.exists(self.conf.task_filename):
